fix: areabrick detection
The way of checking if a page had certain areabricks was too imprecise. It only checked if the brick identifier was in the `data` column of the `documents_editables` table.

However, if the identifier is searched for without a delimiter, it will erroneously be found as a substring in longer identifiers (for example, “buttons” and “shortcut-buttons”).

Therefore, we should only check whether the identifier occurs exactly as a type of an areablock.